+++ packages/components/src/components/telekom/nav-icon/nav-icon.spec.ts
@@ -32,6 +32,7 @@ describe('nav-icon', () => {
                 click-link="true"
                 icon="icon"
                 is-mobile-menu-open="true"
+                mobile-menu-open="true"
                 ref-mobile-menu-toggle="">
               </scale-nav-icon>`,
     });
@@ -41,7 +42,9 @@ describe('nav-icon', () => {
     expect(specPage.rootInstance.href).toBe('href');
     expect(specPage.rootInstance.clickLink).toBe('true');
     expect(specPage.rootInstance.icon).toBe('icon');
+    // DEPRECATED - active should replace isActive
     expect(specPage.rootInstance.isMobileMenuOpen).toBe(true);
+    expect(specPage.rootInstance.mobileMenuOpen).toBe(true);
     expect(specPage.rootInstance.refMobileMenuToggle).toBe('');
   });
   it('should trigger onClick()', async () => {