+++ packages/components/src/components/telekom/nav-icon/nav-icon.tsx
@@ -28,8 +28,14 @@ export class NavIcon {
   @Prop() href?: string = 'javascript:void(0);';
   @Prop() clickLink: any;
   @Prop() icon: string;
+  // DEPRECATED - mobileMenuOpen should replace isMobileMenuOpen
   @Prop() isMobileMenuOpen?: boolean = false;
+  @Prop() mobileMenuOpen?: boolean = false;
   @Prop() refMobileMenuToggle?: any;
+  @Prop() refMobileUserMenuToggle?: any;
+  @Prop() refUserMenuToggle?: any;
+  @Prop() badge: boolean = false;
+  @Prop() badgeLabel: number;
 
   componentWillRender() {
     // make sure the deprecated props overwrite the actual ones if used
@@ -43,6 +49,15 @@ export class NavIcon {
         source: this.host,
       });
     }
+    if (this.isMobileMenuOpen !== false) {
+      statusNote({
+        tag: 'deprecated',
+        message:
+          'Property "isMobileMenuOpen" is deprecated. Please use the "mobileMenuOpen" property!',
+        type: 'warn',
+        source: this.host,
+      });
+    }
   }
 
   render() {
@@ -50,7 +65,11 @@ export class NavIcon {
       <li class={this.getCssClassMap()}>
         <a
           class="meta-navigation__item-link"
-          ref={this.refMobileMenuToggle}
+          ref={
+            this.refMobileMenuToggle ||
+            this.refMobileUserMenuToggle ||
+            this.refUserMenuToggle
+          }
           href={this.href}
           onClick={this.clickLink}
           onKeyDown={(event) => {
@@ -63,10 +82,25 @@ export class NavIcon {
             }
           }}
         >
-          {renderIcon({
-            tag: `scale-icon-${this.icon}`,
-            attributes: { class: 'meta-navigation__item-link-icon' },
-          })}
+          {this.badge || (this.badgeLabel && this.badge) || this.badgeLabel ? (
+            <scale-notification-badge label={this.badgeLabel} type="nav-icon">
+              {renderIcon({
+                tag: `scale-icon-${this.icon}`,
+                attributes: {
+                  class: 'meta-navigation__item-link-icon',
+                  selected: this.active || this.isActive,
+                },
+              })}
+            </scale-notification-badge>
+          ) : (
+            renderIcon({
+              tag: `scale-icon-${this.icon}`,
+              attributes: {
+                class: 'meta-navigation__item-link-icon',
+                selected: this.active || this.isActive,
+              },
+            })
+          )}
           <span class="meta-navigation__item-label">
             <slot></slot>
           </span>
@@ -78,7 +112,10 @@ export class NavIcon {
   getCssClassMap() {
     return classNames(
       'meta-navigation__item',
-      (this.active || this.isActive || this.isMobileMenuOpen) &&
+      (this.active ||
+        this.isActive ||
+        this.mobileMenuOpen ||
+        this.isMobileMenuOpen) &&
         'meta-navigation__item--selected',
       !!this.refMobileMenuToggle && 'mobile-menu'
     );